Add a whole bunch of stuff

+++ Ryu64.MIPS/Interpreter/InstInterpJumpBranch.cs
@@ -10,14 +10,14 @@ namespace Ryu64.MIPS
         {
             Registers.R4300.PC += 4;
             R4300.InterpretOpcode(R4300.memory.ReadUInt32(Registers.R4300.PC));
-            if (Registers.R4300.Reg[Desc.op1] == Registers.R4300.Reg[Desc.op2])
+            if ((long)Registers.R4300.Reg[Desc.op1] == (long)Registers.R4300.Reg[Desc.op2])
                 Registers.R4300.PC += (uint)((short)(Desc.Imm - 1) << 2);
         }
 
         public static void BEQL(OpcodeTable.OpcodeDesc Desc)
         {
             Registers.R4300.PC += 4;
-            if (Registers.R4300.Reg[Desc.op1] == Registers.R4300.Reg[Desc.op2])
+            if ((long)Registers.R4300.Reg[Desc.op1] == (long)Registers.R4300.Reg[Desc.op2])
             {
                 R4300.InterpretOpcode(R4300.memory.ReadUInt32(Registers.R4300.PC));
                 Registers.R4300.PC += (uint)((short)(Desc.Imm - 1) << 2);
@@ -29,7 +29,7 @@ namespace Ryu64.MIPS
         {
             Registers.R4300.PC += 4;
             R4300.InterpretOpcode(R4300.memory.ReadUInt32(Registers.R4300.PC));
-            if (Registers.R4300.Reg[Desc.op1] >= 0)
+            if ((long)Registers.R4300.Reg[Desc.op1] >= 0)
                 Registers.R4300.PC += (uint)((short)(Desc.Imm - 1) << 2);
         }
 
@@ -38,14 +38,14 @@ namespace Ryu64.MIPS
             Registers.R4300.PC += 4;
             R4300.InterpretOpcode(R4300.memory.ReadUInt32(Registers.R4300.PC));
             Registers.R4300.Reg[31] = Registers.R4300.PC;
-            if (Registers.R4300.Reg[Desc.op1] >= 0)
+            if ((long)Registers.R4300.Reg[Desc.op1] >= 0)
                 Registers.R4300.PC += (uint)((short)(Desc.Imm - 1) << 2);
         }
 
         public static void BLEZL(OpcodeTable.OpcodeDesc Desc)
         {
             Registers.R4300.PC += 4;
-            if (Registers.R4300.Reg[Desc.op1] <= 0)
+            if ((long)Registers.R4300.Reg[Desc.op1] <= 0)
             {
                 R4300.InterpretOpcode(R4300.memory.ReadUInt32(Registers.R4300.PC));
                 Registers.R4300.PC += (uint)((short)(Desc.Imm - 1) << 2);
@@ -57,7 +57,7 @@ namespace Ryu64.MIPS
         {
             Registers.R4300.PC += 4;
             R4300.InterpretOpcode(R4300.memory.ReadUInt32(Registers.R4300.PC));
-            if (Registers.R4300.Reg[Desc.op1] <= 0)
+            if ((long)Registers.R4300.Reg[Desc.op1] <= 0)
                 Registers.R4300.PC += (uint)((short)(Desc.Imm - 1) << 2);
         }
 
@@ -66,7 +66,7 @@ namespace Ryu64.MIPS
             Registers.R4300.PC += 4;
             R4300.InterpretOpcode(R4300.memory.ReadUInt32(Registers.R4300.PC));
             Registers.R4300.Reg[31] = Registers.R4300.PC;
-            if (Registers.R4300.Reg[Desc.op1] <= 0)
+            if ((long)Registers.R4300.Reg[Desc.op1] <= 0)
                 Registers.R4300.PC += (uint)((short)(Desc.Imm - 1) << 2);
         }
 
@@ -74,14 +74,14 @@ namespace Ryu64.MIPS
         {
             Registers.R4300.PC += 4;
             R4300.InterpretOpcode(R4300.memory.ReadUInt32(Registers.R4300.PC));
-            if (Registers.R4300.Reg[Desc.op1] != Registers.R4300.Reg[Desc.op2])
+            if ((long)Registers.R4300.Reg[Desc.op1] != (long)Registers.R4300.Reg[Desc.op2])
                 Registers.R4300.PC += (uint)((short)(Desc.Imm - 1) << 2);
         }
 
         public static void BNEL(OpcodeTable.OpcodeDesc Desc)
         {
             Registers.R4300.PC += 4;
-            if (Registers.R4300.Reg[Desc.op1] != Registers.R4300.Reg[Desc.op2])
+            if ((long)Registers.R4300.Reg[Desc.op1] != (long)Registers.R4300.Reg[Desc.op2])
             {
                 R4300.InterpretOpcode(R4300.memory.ReadUInt32(Registers.R4300.PC));
                 Registers.R4300.PC += (uint)((short)(Desc.Imm - 1) << 2);
